refactor(x): one conversation-item renderer inside ChatSessionPane
The two renderConversationItem copies (App's rich variant, the
sidebar's plainer one) merge into ChatSessionPane; the renderItem
injection is gone. Genuine differences became flagged props —
richToolCards (App-only CodingRunBlock/SubAgentBlock/AppActionCard vs
generic Tool rows), detailedToolErrors (sidebar surfaced real tool
error text), optional onCodePermissionResponse / onComposioConnected —
never silently changed. Shared markdown constants move to
lib/markdown-render.ts (non-component module). One deliberate
unification: both AutoScrollPre copies now use App's useLayoutEffect
timing. Net -200 lines (App -246, sidebar -215, chat-session +250,
lib +11).
